@@ -42,171 +45,187 @@
   },
   "contributors": [
     {
-      "name": "Hugh Kennedy",
-      "email": "hughskennedy@gmail.com"
+      "email": "tobias.koppers@googlemail.com",
+      "name": "Tobias Koppers"
     },
     {
-      "name": "Tobias Koppers",
-      "email": "tobias.koppers@googlemail.com"
+      "email": "duncan@dweebd.com",
+      "name": "Duncan Beevers"
     },
     {
-      "name": "Stephen Crane",
-      "email": "scrane@mozilla.com"
+      "email": "scrane@mozilla.com",
+      "name": "Stephen Crane"
     },
     {
-      "name": "Ryan Seddon",
-      "email": "seddon.ryan@gmail.com"
+      "email": "seddon.ryan@gmail.com",
+      "name": "Ryan Seddon"
     },
     {
-      "name": "Miles Elam",
-      "email": "miles.elam@deem.com"
+      "email": "miles.elam@deem.com",
+      "name": "Miles Elam"
     },
     {
-      "name": "Mihai Bazon",
-      "email": "mihai.bazon@gmail.com"
+      "email": "mihai.bazon@gmail.com",
+      "name": "Mihai Bazon"
     },
     {
-      "name": "Michael Ficarra",
-      "email": "github.public.email@michael.ficarra.me"
+      "email": "github.public.email@michael.ficarra.me",
+      "name": "Michael Ficarra"
     },
     {
-      "name": "Todd Wolfson",
-      "email": "todd@twolfson.com"
+      "email": "todd@twolfson.com",
+      "name": "Todd Wolfson"
     },
     {
-      "name": "Alexander Solovyov",
-      "email": "alexander@solovyov.net"
+      "email": "alexander@solovyov.net",
+      "name": "Alexander Solovyov"
     },
     {
-      "name": "Felix Gnass",
-      "email": "fgnass@gmail.com"
+      "email": "fgnass@gmail.com",
+      "name": "Felix Gnass"
     },
     {
-      "name": "Conrad Irwin",
-      "email": "conrad.irwin@gmail.com"
+      "email": "conrad.irwin@gmail.com",
+      "name": "Conrad Irwin"
     },
     {
-      "name": "usrbincc",
-      "email": "usrbincc@yahoo.com"
+      "email": "usrbincc@yahoo.com",
+      "name": "usrbincc"
     },
     {
-      "name": "David Glasser",
-      "email": "glasser@davidglasser.net"
+      "email": "glasser@davidglasser.net",
+      "name": "David Glasser"
     },
     {
-      "name": "Chase Douglas",
-      "email": "chase@newrelic.com"
+      "email": "chase@newrelic.com",
+      "name": "Chase Douglas"
     },
     {
-      "name": "Evan Wallace",
-      "email": "evan.exe@gmail.com"
+      "email": "evan.exe@gmail.com",
+      "name": "Evan Wallace"
     },
     {
-      "name": "Heather Arthur",
-      "email": "fayearthur@gmail.com"
+      "email": "fayearthur@gmail.com",
+      "name": "Heather Arthur"
     },
     {
-      "name": "Duncan Beevers",
-      "email": "duncan@dweebd.com"
+      "email": "hughskennedy@gmail.com",
+      "name": "Hugh Kennedy"
     },
     {
-      "name": "David Glasser",
-      "email": "glasser@davidglasser.net"
+      "email": "glasser@davidglasser.net",
+      "name": "David Glasser"
     },
     {
-      "name": "Simon Lydell",
-      "email": "simon.lydell@gmail.com"
+      "email": "simon.lydell@gmail.com",
+      "name": "Simon Lydell"
     },
     {
-      "name": "Jmeas Smith",
-      "email": "jellyes2@gmail.com"
+      "email": "jellyes2@gmail.com",
+      "name": "Jmeas Smith"
     },
     {
-      "name": "Michael Z Goddard",
-      "email": "mzgoddard@gmail.com"
+      "email": "mzgoddard@gmail.com",
+      "name": "Michael Z Goddard"
     },
     {
-      "name": "azu",
-      "email": "azu@users.noreply.github.com"
+      "email": "azu@users.noreply.github.com",
+      "name": "azu"
     },
     {
-      "name": "John Gozde",
-      "email": "john@gozde.ca"
+      "email": "john@gozde.ca",
+      "name": "John Gozde"
     },
     {
-      "name": "Adam Kirkton",
-      "email": "akirkton@truefitinnovation.com"
+      "email": "akirkton@truefitinnovation.com",
+      "name": "Adam Kirkton"
     },
     {
-      "name": "Chris Montgomery",
-      "email": "christopher.montgomery@dowjones.com"
+      "email": "christopher.montgomery@dowjones.com",
+      "name": "Chris Montgomery"
     },
     {
-      "name": "J. Ryan Stinnett",
-      "email": "jryans@gmail.com"
+      "email": "jryans@gmail.com",
+      "name": "J. Ryan Stinnett"
     },
     {
-      "name": "Jack Herrington",
-      "email": "jherrington@walmartlabs.com"
+      "email": "jherrington@walmartlabs.com",
+      "name": "Jack Herrington"
     },
     {
-      "name": "Chris Truter",
-      "email": "jeffpalentine@gmail.com"
+      "email": "jeffpalentine@gmail.com",
+      "name": "Chris Truter"
     },
     {
-      "name": "Daniel Espeset",
-      "email": "daniel@danielespeset.com"
+      "email": "daniel@danielespeset.com",
+      "name": "Daniel Espeset"
     },
     {
-      "name": "Jamie Wong",
-      "email": "jamie.lf.wong@gmail.com"
+      "email": "jamie.lf.wong@gmail.com",
+      "name": "Jamie Wong"
     },
     {
-      "name": "Eddy Bruël",
-      "email": "ejpbruel@mozilla.com"
+      "email": "ejpbruel@mozilla.com",
+      "name": "Eddy Bruël"
     },
     {
-      "name": "Hawken Rives",
-      "email": "hawkrives@gmail.com"
+      "email": "hawkrives@gmail.com",
+      "name": "Hawken Rives"
+    },
+    {
+      "email": "giladp007@gmail.com",
+      "name": "Gilad Peleg"
+    },
+    {
+      "email": "djchie.dev@gmail.com",
+      "name": "djchie"
+    },
+    {
+      "email": "garysye@gmail.com",
+      "name": "Gary Ye"
+    },
+    {
+      "email": "nicolas.lalevee@hibnet.org",
+      "name": "Nicolas Lalevée"
     }
   ],
-  "dependencies": {
-    "amdefine": ">=0.0.4"
-  },
+  "dependencies": {},
   "description": "Generates and consumes source maps",
   "devDependencies": {
-    "dryice": ">=0.4.8"
+    "doctoc": "^0.15.0",
+    "webpack": "^1.12.0"
   },
-  "directories": {
-    "lib": "./lib"
-  },
+  "directories": {},
   "dist": {
-    "shasum": "dc9f3114394ab7c1f9782972f3d11820fff06f1f",
-    "tarball": "http://registry.npmjs.org/source-map/-/source-map-0.4.2.tgz"
+    "shasum": "82674b85a71b0be76c3e7416d15e9f5252eb3be0",
+    "tarball": "http://registry.npmjs.org/source-map/-/source-map-0.5.3.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=0.10.0"
   },
-  "homepage": "https://github.com/mozilla/source-map",
-  "licenses": [
-    {
-      "type": "BSD",
-      "url": "http://opensource.org/licenses/BSD-3-Clause"
-    }
+  "files": [
+    "source-map.js",
+    "lib/",
+    "dist/source-map.debug.js",
+    "dist/source-map.js",
+    "dist/source-map.min.js",
+    "dist/source-map.min.js.map"
   ],
-  "main": "./lib/source-map.js",
+  "homepage": "https://github.com/mozilla/source-map",
+  "license": "BSD-3-Clause",
+  "main": "./source-map.js",
   "maintainers": [
     {
-      "name": "mozilla-devtools",
-      "email": "mozilla-developer-tools@googlegroups.com"
+      "email": "mozilla-developer-tools@googlegroups.com",
+      "name": "mozilla-devtools"
     },
     {
-      "name": "mozilla",
-      "email": "dherman@mozilla.com"
+      "email": "dherman@mozilla.com",
+      "name": "mozilla"
     },
     {
-      "name": "nickfitzgerald",
-      "email": "fitzgen@gmail.com"
+      "email": "fitzgen@gmail.com",
+      "name": "nickfitzgerald"
     }
   ],
   "name": "source-map",
